Faculty Part/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0

# LLM & Embeddings (using Ollama - no API keys needed)
langchain==0.1.4

# Vector Database
qdrant-client==1.7.3

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.17.2

# Search & Ranking
rank-bm25==0.2.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Utilities
python-dotenv==1.0.1
pydantic-settings==2.1.0

# Additional dependencies
sentence-transformers==2.3.1
pdf2image==1.16.3
